pc - update DeveloperPage story for storybook 8, and reorganize CourseOverTime stories to be together in same folder
